Keep split loader tests after rebase
